requirements.txt
pika~=1.3.2 Tương tác với RabbitMQ, 1.3.2 tương thích với tất cả phiên bản RabbitMQ
torch Deep Learning framework, 2.7.0
torchvision Deep Learning framework, 0.22.0
numpy Xử lý mảng số học, 2.2.5
requests~=2.32.3 Gửi HTTP requests, 2.32.3
pyyaml Đọc/ghi file YAML, 6.0.2
opencv-python Xử lý ảnh và video, 4.11.0.86
ultralytics Framework train model YOLO, 8.3.122




